requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
Flask==3.1.1
Flask-Caching==2.3.1
flask-core==2.9.0
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.2
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
nltk==3.9.1
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.2.0
regex==2024.11.6
requests==2.32.3
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.41
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wordcloud==1.9.4
